% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnglitchr-package.R
\docType{package}
\name{pnglitchr-package}
\alias{pnglitchr}
\alias{pnglitchr-package}
\title{pnglitchr: PNG Glitching in R}
\description{
Offers a thin wrapper around \url{https://github.com/chikoski/png-glitch}, a library to glitch PNG images.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Akiru Kato \email{paithiov909@gmail.com}
}

Other contributors:
\itemize{
  \item N. Shimizu (Author of the original 'png-glitch' library) [copyright holder]
}

}
\keyword{internal}
