% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnglitchr-package.R
\name{pnglitch}
\alias{pnglitch}
\alias{count_scanlines}
\alias{glitch_replace}
\alias{glitch_remove}
\alias{glitch_apply}
\title{Create a glitched PNG image}
\usage{
count_scanlines(x)

glitch_replace(x, times)

glitch_remove(x, from, to)

glitch_apply(
  x,
  from,
  to,
  filter_type = c("none", "sub", "up", "average", "paeth")
)
}
\arguments{
\item{x}{A character string specifying the path to a PNG file
or a raw vector containing a PNG image data.}

\item{times}{An integer specifying the number of times to copy.}

\item{from, to}{Scan line index.}

\item{filter_type}{Filter type.
One of "none", "sub", "up", "average", "paeth".}
}
\value{
A raw vector containing a glitched PNG image data.
}
\description{
Create a glitched PNG image
}
\details{
The following functions are available:
\itemize{
\item \code{count_scanlines()}: Returns the total number of scan lines in the PNG image.
\item \code{glitch_replace()}: Replaces the scan lines randomly.
\item \code{glitch_remove()}: Removes filter from the scan lines.
\item \code{glitch_apply()}: Applies a specified filter to the scan lines.
}
}
