% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnglitchr-package.R
\docType{package}
\name{pnglitchr-package}
\alias{pnglitchr}
\alias{pnglitchr-package}
\title{pnglitchr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item N. Shimizu [copyright holder]
}

}
\keyword{internal}
